Fix assigning Flowing Water (regression from 6b00a3430d080ecf0b47ed0e13be535b7f2d5eef)
when rendering WITH water (-nowater not set) the assignBlock water case did not hit
and ended in the default else tree, which was only supposed to happen for doublesteps.
Flowing water has a "col" value, depending on its height, we did hit the
sandstone & wood & cobblestone doublestep case, which is why only height flowing water
was replaced and the "low" flowing water not.