All variants now supported
In the colors array, colors[block][VINDEX] can be used to find the start
of the variant list.
If colors[SAND][VINDEX] equals 23, sand variants can be found at colors[255+23].
This is set only in the 256 first blocks, as every block after that is a
variant.

In the variant list, the value at VINDEX is set only at the first entry
of the variant list, and serves as a variant count. There are 4 types of
logs, but they can be placed on different axis, and their 'variant'
value can therefore be higher than 4. The variant count is there to be
able to know when the variant is actually another value.

You can also set the way the block in displayed by setting
colors[BLOCK_TYPE] with a value of the enum in Blocks::BlockTypes.
Not setting it will make it be displayed as a full block.