Massive code restructuring
Removed OrientedMap for enhanced coordinates.
Moved pixel buffer from the Image class to the IsometricCanvas to simplify
buffer merging once the threads come along.

The canvas is as key as ever in making a transition between a list of chunks
and a 2D pixel matrix. It contains all the drawing code, and the
`render` function that stuck like a sore thumb in main.

No more shady bound swapping at initialization, and index swapping at
runtime: the canvas handles it all.